processing_worker/requirements.txt
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
psycopg2-binary==2.9.9
